# docs/openapi-client.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Guardian Client API",
    "description": "Client-facing Guardian HTTP API consumed by SDKs and packages.",
    "license": {
      "name": "AGPL-3.0",
      "identifier": "AGPL-3.0"
    },
    "version": "0.17.0-rc.3"
  },
  "paths": {
...
    "/delta/proposal": {
      "get": {
        "tags": [
          "client"
        ],
        "summary": "List all in-flight multisig proposals for an account.",
        "operationId": "get_delta_proposals",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending proposals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or replay rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the configured size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-pubkey": [],
            "x-signature": [],
            "x-timestamp": []
          }
        ]
      },
      "put": {
        "tags": [
          "client"
        ],
        "summary": "Add a cosigner signature to an existing multisig proposal. When the\nsignature threshold is reached the proposal is promoted to a delta.",
        "operationId": "sign_delta_proposal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signature accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeltaObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or replay rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Proposal not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Proposal already signed by this signer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the configured size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-pubkey": [],
            "x-signature": [],
            "x-timestamp": []
          }
        ]
      },
      "post": {
        "tags": [
          "client"
        ],
        "summary": "Create a new multisig delta proposal for cosigners to sign.",
        "operationId": "push_delta_proposal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeltaProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Proposal created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeltaProposalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid proposal payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed or replay rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflicting / too many pending proposals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the configured size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-pubkey": [],
            "x-signature": [],
            "x-timestamp": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorMeta": {
        "type": "object",
        "description": "Structured machine-readable side-data on the error wire object\n(feature `009-human-readable-errors`). `retryable` is always present;\nthe rest appear only for the codes that carry them.",
        "required": [
          "retryable"
        ],
        "properties": {
          "missing_permissions": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Lex-sorted permissions the operator lacks. Present only for\n`GUARDIAN_INSUFFICIENT_OPERATOR_PERMISSION`."
          },
          "paused_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC 3339 pause timestamp. Present only for `GUARDIAN_ACCOUNT_PAUSED`."
          },
          "paused_reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pause reason. Present only for `GUARDIAN_ACCOUNT_PAUSED`."
          },
          "released_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC 3339 timestamp of the guardian-switch release. Present only\nfor `GUARDIAN_ACCOUNT_RELEASED`."
          },
          "retry_after_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Seconds to wait before retrying. Present only for\n`rate_limit_exceeded` (the `Retry-After` header carries the same value).",
            "minimum": 0
          },
          "retryable": {
            "type": "boolean",
            "description": "Whether retrying the same request could plausibly succeed."
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "Wire shape of a Guardian error response body: `{ code, message, meta }`\n(feature `009-human-readable-errors`). Mirrors the object produced by\n[`crate::error::GuardianError`]'s `IntoResponse` impl and carried\nidentically on the gRPC `Status.details`. The legacy `success`/`error`\nfields are gone; the diagnostic detail is logged server-side only.\nDocumented as the body of every non-2xx response.",
        "required": [
          "code",
          "message",
          "meta"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code (e.g. `account_not_found`). The\nclient branch + i18n key. Branch on this, never on `message`."
          },
          "message": {
            "type": "string",
            "description": "Short, user-safe message; safe to display verbatim. Wording is not\npart of the stable contract."
          },
          "meta": {
            "$ref": "#/components/schemas/ApiErrorMeta",
            "description": "Structured machine-readable side-data."
          }
        }
      },
...
      "DeltaMetadata": {
        "type": "object",
        "description": "Persisted activity metadata for a canonical delta. Stored as JSONB\nin the `deltas.metadata` column. `None` for EVM deltas and any\nhistorical row never reprocessed by [`build_metadata`].",
        "required": [
          "category"
        ],
        "properties": {
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetSummary"
            },
            "description": "Assets surfaced in deterministic order. Empty when the\ntransaction does not move an asset or extraction failed.\nMulti-asset transactions populate every extractable entry so\nclients do not show a misleading single-asset summary."
          },
          "category": {
            "$ref": "#/components/schemas/DashboardDeltaCategory"
          },
          "counterparty": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CounterpartySummary",
                "description": "Counterparty of the transaction. `None` for transactions\nwithout a clear sender/recipient (admin ops, swaps, etc.)."
              }
            ]
          },
          "note_counts": {
            "$ref": "#/components/schemas/NoteCounts"
          },
          "proposal": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProposalMetadata",
                "description": "Multisig proposal intent lifted from the matching\n`delta_proposals` row at push time. Absent for single-key\n`push_delta`, EVM deltas, and pushes where no proposal matched."
              }
            ]
          }
        }
      },
      "DeltaObject": {
        "type": "object",
        "description": "Delta object",
        "required": [
          "account_id",
          "nonce",
          "prev_commitment",
          "delta_payload",
          "ack_sig",
          "ack_pubkey",
          "ack_scheme",
          "status"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "ack_pubkey": {
            "type": "string"
          },
          "ack_scheme": {
            "type": "string"
          },
          "ack_sig": {
            "type": "string"
          },
          "delta_payload": {
            "type": "object",
            "description": "Opaque, schema-free JSON payload describing the state delta."
          },
          "metadata": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeltaMetadata",
                "description": "Typed dashboard metadata derived at push time. Stored as JSONB\nin the `deltas.metadata` column. `None` for EVM deltas and any\nhistorical row never reprocessed by the push-time pipeline."
              }
            ]
          },
          "new_commitment": {
            "type": [
              "string",
              "null"
            ]
          },
          "nonce": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "prev_commitment": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DeltaStatus"
          }
        }
      },
      "DeltaProposalRequest": {
        "type": "object",
        "required": [
          "account_id",
          "nonce",
          "delta_payload"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "delta_payload": {
            "type": "object",
            "description": "Opaque, schema-free multisig proposal payload."
          },
          "nonce": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "DeltaProposalResponse": {
        "type": "object",
        "required": [
          "delta",
          "commitment"
        ],
        "properties": {
          "commitment": {
            "type": "string"
          },
          "delta": {
            "$ref": "#/components/schemas/DeltaObject"
          }
        }
      },
...
      "ProposalSignature": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "signature",
              "scheme"
            ],
            "properties": {
              "scheme": {
                "type": "string",
                "enum": [
                  "falcon"
                ]
              },
              "signature": {
                "type": "string",
                "description": "Hex-encoded Falcon signature"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "signature",
              "scheme"
            ],
            "properties": {
              "public_key": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Hex-encoded ECDSA public key (required for signature preparation)"
              },
              "scheme": {
                "type": "string",
                "enum": [
                  "ecdsa"
                ]
              },
              "signature": {
                "type": "string",
                "description": "Hex-encoded ECDSA secp256k1 signature"
              }
            }
          }
        ],
        "description": "Signature type for delta proposals"
      },
      "ProposalsResponse": {
        "type": "object",
        "required": [
          "proposals"
        ],
        "properties": {
          "proposals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeltaObject"
            }
          }
        }
      },
...
      "SignProposalRequest": {
        "type": "object",
        "required": [
          "account_id",
          "commitment",
          "signature"
        ],
        "properties": {
          "account_id": {
            "type": "string"
          },
          "commitment": {
            "type": "string"
          },
          "signature": {
            "$ref": "#/components/schemas/ProposalSignature"
          }
        }
      },
...
    },
    "securitySchemes": {
      "x-pubkey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-pubkey"
      },
      "x-signature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-signature"
      },
      "x-timestamp": {
        "type": "apiKey",
        "in": "header",
        "name": "x-timestamp"
      }
    }
  },
...
}